requirements.txt
feedparser==6.0.11
jinja2==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
